+++ src/jauge.py
@@ -1,5 +1,8 @@
-from random import *
+from random import randint
 from constante import *
+from popularite import *
+from justice import *
+from legalite import *
 
 class Jauge:
 
@@ -41,122 +44,3 @@ class Jauge:
         if pop >= 4000000000 :                                          #arrivé au grace maximal, président des nations, fin du jeu.
             Liste_evenement.set_grade = PRESIDENT_DES_NATIONS
             return Gestion.fin_jeu(1)
-
-
-#Popularité hérite de Jauge
-
-class Popularite(Jauge):
-
-    def __init__(self, popularite: int = 10) -> None:
-        super().__init__()
-        self.popularite = 10 
-
-
-    def set_pop(self, v:int) -> int :
-        """setter popularite
-
-        Args:
-            v (int): popularite voulue
-
-        Returns:
-            int: popularite
-        """
-        self.popularite = v
-        return self.popularite
-
-    def get_pop(self):
-        """getter popularité"""
-        return self.popularite
-
-    def add_pop(self, v:int) -> int : 
-        """ajoute de la popularite    
-
-        Args:
-            v (int): valeur qui va être ajoutée
-
-        Returns:
-            int: popularité
-        """        
-        if type(v) == int :
-            self.popularite = self.popularite + v
-            return self.popularite
-
-#Justice hérite de Jauge
-
-class Justice(Jauge):
-
-    def __init__(self, justice: float = 0) -> None:
-        super().__init__()
-
-    def set_jus(self, v:int) -> int :
-        """setter justice
-
-        Args:
-            v (int): pourcentage de jusitce voulue
-
-        Returns:
-            int: justice
-        """
-        self.justice = v
-        return self.justice
-
-    def get_jus(self):
-        """getter justice"""
-        return self.justice
-
-    def add_jus(self, v:int):
-        """ ajoute de la justice
-
-        Args:
-            v (int): valeur qui va être ajoutée 
-
-        Returns:
-            int : justice
-        """
-        self.justice = self.justice + v
-        if self.justice > 100 :             #pour ne pas que la justice soit au dessus de 100%
-            self.justice = 100
-        if self.justice < 0 :               #et pas en dessous de 0%
-            self.justice = 0
-        else :
-            return self.justice
-
-#Legalite qui hérite de Jauge
-
-class Legalite(Jauge):
-
-    def __init__(self, legalite : float = 100) -> None:
-        super().__init__()
-
-    def set_leg(self, v:int) -> int :
-        """setter legalité
-
-        Args:
-            v (int): légalité voulue
-
-        Returns:
-            int: légalité
-        """
-        self.legalite = v
-        return self.legalite
-
-    def get_leg(self):
-        """getter legalite"""
-        return self.legalite
-
-    def add_leg(self, v):
-        """ajoute de la légalité
-
-        Args:
-            v (int): valeur ajoutée
-
-        Returns:
-            int : legalité
-        """
-        self.legalite = self.legalite + v
-        if self.legalite > 100 :             #pour ne pas que la legalite soit au dessus de 100%        
-            self.legalite = 100
-        if self.legalite < 0 :               #et pas en dessous de 0%
-            self.legalite = 0
-        else :
-            return self.legalite
